% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PadeApproximation.R
\name{approximateLikelihoodUsingPade}
\alias{approximateLikelihoodUsingPade}
\title{Approximate Cox regression using Pade approximation}
\usage{
approximateLikelihoodUsingPade(survivalData, bounds = c(log(0.1), log(10)))
}
\arguments{
\item{survivalData}{A data frame with the person-level survival data. See details.}

\item{bounds}{The bounds on the effect size used to fit the approximation.}
}
\value{
A data frame with the parameters of the Pade approximation.
}
\description{
Approximate Cox regression using Pade approximation
}
\details{
The \code{survivalData} data frame should have the following columns:
\itemize{
\item \strong{x}: the binary exposure variable.
\item \strong{time}: the time to either the outcome or censoring.
\item \strong{y}: indicator whether at the end of \code{time} the outcome occurred (1) or censoring (0).
\item \strong{stratumId}: an integer indicating the stratum (e.g. propensity-score strata).
}
}
