% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodApproximation.R
\name{buildLabelReferences}
\alias{buildLabelReferences}
\title{Build a list of references that map likelihood names to integer labels for later use}
\usage{
buildLabelReferences(data)
}
\arguments{
\item{data}{The likelihood data. Can be a single approximation, approximations
from multiple sites, or (adaptive) grid profile likelihoods.}
}
\description{
Build a list of references that map likelihood names to integer labels for later use
}
\examples{
data("likelihoodProfileLists")
refLabs = buildLabelReferences(likelihoodProfileLists)

}
