% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalMetaAnalysis.R
\name{computeHierarchicalMetaAnalysis}
\alias{computeHierarchicalMetaAnalysis}
\title{Compute a Bayesian random-effects hierarchical meta-analysis}
\usage{
computeHierarchicalMetaAnalysis(
  data,
  chainLength = 1100000,
  burnIn = 1e+05,
  subSampleFrequency = 100,
  alpha = 0.05,
  effectPriorStd = 1,
  globalExposureEffectPrior = c(0, 2),
  effectPrecisionPrior = c(1, 1),
  errorPrecisionPrior = c(1, 1),
  errorPrecisionStartValue = 1,
  includeSourceEffect = TRUE,
  seed = 1
)
}
\arguments{
\item{data}{A data frame containing either normal, skew-normal, custom parametric,
or grid likelihood data, with one row per database.}

\item{chainLength}{Number of MCMC iterations.}

\item{burnIn}{Number of MCMC iterations to consider as burn in.}

\item{subSampleFrequency}{Subsample frequency for the MCMC.}

\item{alpha}{The alpha (expected type I error) used for the credible intervals.}

\item{seed}{The seed for the random number generator.}

\item{priorSd}{A two-dimensional vector with the standard deviation of the prior for mu
and tau, respectively.}

\item{robust}{Whether or not to use a t-distribution model; default: FALSE.}

\item{df}{Degrees of freedom for the t-model, only used if robust is TRUE.}
}
\value{
A data frame with the point estimates and 95\% credible intervals for the mu and tau parameters (the
mean and standard deviation of the distribution from which the per-site effect sizes are drawn).
Attributes of the data frame contain the MCMC trace and the detected approximation type.
}
\description{
Compute a Bayesian hierarchical meta-analysis (two-level model) using the Markov chain Monte Carlo (MCMC) engine BEAST.
A normal prior is used for the global effects and primary / secondary effects;
a half normal prior is used for the standard deviation; a gamma prior is used for the precision parameters.
}
\seealso{
\link{approximateLikelihood}, \link{computeFixedEffectMetaAnalysis}
}
