% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalMetaAnalysis.R
\name{computeSourceSpecificEffects2}
\alias{computeSourceSpecificEffects2}
\title{Compute source-specific biases and bias-corrected estimates from hierarchical meta analysis results, version 2.0}
\usage{
computeSourceSpecificEffects2(
  estimates,
  ooiLPs,
  useDatabaseName = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{estimates}{A dataframe as output from the \code{computeHierarchicalMetaAnalysis()} function.}

\item{ooiLPs}{Profile likehoods of the outcome(s) of interest; either a list with \code{database_id} as name, or a list of such lists.}

\item{useDatabaseName}{Use name of database (\code{database_id}) as source id in results dataframe? Default is FALSE.}

\item{alpha}{The alpha (expected type I error) used for the credible intervals.}
}
\value{
A data frame with point estimates, (1-alpha) credible intervals and sample standard errors for the
effect size after bias correction within each data source.
}
\description{
Compute source-specific biases and obtain bias-corrected estimates for each data source,
given the results from \code{computeHierarchicalMetaAnalysis()}.
}
\seealso{
\link{computeHierarchicalMetaAnalysis}
}
