% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodApproximation.R
\name{constructDataModel}
\alias{constructDataModel}
\title{Construct \code{DataModel} objects from approximate likelihood or profile likelihood data}
\usage{
constructDataModel(data, labelReferences = NULL)
}
\arguments{
\item{data}{The likelihood data. Can be a single approximation, approximations
from multiple sites, or (adaptive) gride profile likelihoods.}

\item{labelReferences}{Optional parameter that provides a reference list that
maps string names to integer indices; only applies to
"grid" or "adaptive grip" type of data.}
}
\description{
Construct \code{DataModel} objects from approximate likelihood or profile likelihood data
}
\examples{
data("likelihoodProfileLists")
dataModel <- constructDataModel(likelihoodProfileLists[[1]])

}
