% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createApproximations}
\alias{createApproximations}
\title{Create likelihood approximations from individual-trajectory data}
\usage{
createApproximations(populations, approximation)
}
\arguments{
\item{populations}{Individual-level population data}

\item{approximation}{Type of approximation method}
}
\description{
Create likelihood approximations from individual-trajectory data
}
