% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodApproximation.R
\name{customFunction}
\alias{customFunction}
\title{A custom function to approximate a log likelihood function}
\usage{
customFunction(x, mu, sigma, gamma)
}
\arguments{
\item{x}{The log(hazard ratio) for which to approximate the log likelihood.}

\item{mu}{The position parameter.}

\item{sigma}{The scale parameter.}

\item{gamma}{The skew parameter.}
}
\value{
The approximate log likelihood for the given x.
}
\description{
A custom function to approximate a log likelihood function
}
\examples{
customFunction(x = 0:3, mu = 0, sigma = 1, gamma = 0)
# [1]  0.0 -0.5 -2.0 -4.5

}
