% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{estimateByPooling}
\alias{estimateByPooling}
\title{Estimate effect by pooling data}
\usage{
estimateByPooling(simulation, useCyclops = TRUE)
}
\arguments{
\item{simulation}{An object of type \code{simulation} as created by the 
\code{\link{runSimulation}} function.}

\item{useCyclops}{Should Cyclops be used? Cyclops doesn't assume a normal distribution for the 
likelihood function, and uses likelihood profiling instead.}
}
\value{
A data frame with one row containing the point estimate and confidence interval.
}
\description{
Estimate the hazard ratio across site by pooling the data.
}
