% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{estimateUsingStandardMetaAnalysis}
\alias{estimateUsingStandardMetaAnalysis}
\title{Estimate effect by using standard meta-analysis}
\usage{
estimateUsingStandardMetaAnalysis(simulation, assumeRandomEffect = TRUE,
  useCyclops = FALSE)
}
\arguments{
\item{simulation}{An object of type \code{simulation} as created by the 
\code{\link{runSimulation}} function.}

\item{assumeRandomEffect}{Should meta-analysis for random effects be used?}

\item{useCyclops}{Should Cyclops be used? Cyclops doesn't assume a normal distribution for the 
likelihood function, and uses likelihood profiling instead.}
}
\value{
A data frame with one row containing the point estimate and confidence interval.
}
\description{
Estimate the effect size by first estimating the effect size at each site, and then combine using
standard meta-analysis techniques.
}
