% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{evaluateMetaAnalysisApproach}
\alias{evaluateMetaAnalysisApproach}
\title{Evaluate a meta-analysis approach}
\usage{
evaluateMetaAnalysisApproach(simulationSettings, iterations = 100,
  metaAnalysisFunction, seed = 0, ...)
}
\arguments{
\item{simulationSettings}{An object of type \code{simulationSettings}, created by the 
\code{\link{createSimulationSettings}} function.}

\item{iterations}{Number of times the simulation should be run.}

\item{metaAnalysisFunction}{The function performing the meta-analysis.}

\item{seed}{The random seed to be used. Set before the first iteration.}

\item{...}{Additional arguments that will be passed on to the meta-
analysis function.}
}
\value{
Something
}
\description{
Evaluate a meta-analysis approach
}
\details{
The meta-analysis function should accept a \code{simulation} object, and produce an estimate data frame.
}
