% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalMetaAnalysis.R
\name{extractSourceSpecificEffects}
\alias{extractSourceSpecificEffects}
\title{Compute source-specific biases and bias-corrected estimates from hierarchical meta analysis results}
\usage{
extractSourceSpecificEffects(estimates, alpha = 0.05)
}
\arguments{
\item{estimates}{A data frame as output from the \code{computeHierarchicalMetaAnalysis()} function.}

\item{alpha}{The alpha (expected type I error) used for the credible intervals.}
}
\value{
A data frame with point estimates, 95\% credible intervals and sample standard errors for the
effect size after bias correction within each data source.
}
\description{
Extract source-specific biases and obtain bias-corrected estimates for each data source,
given the results from \code{computeHierarchicalMetaAnalysis()}.
}
\seealso{
\link{computeHierarchicalMetaAnalysis}
}
