% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-data.R
\docType{data}
\name{hmaLikelihoodList}
\alias{hmaLikelihoodList}
\title{Example profile likelihoods for hierarchical meta analysis with bias correction}
\format{
An objects of class \code{list}; the list contains 3 dataframes,
where each dataframe includes four rows of likelihood function values
corresponding to the points in the column names.
}
\usage{
hmaLikelihoodList
}
\description{
A list that contains profile likelihoods for two negative control outcomes and
a synthetic outcome of interest, across four data sources.
Each element of the list contains profile likelihoods for one outcome,
where each row provides profile likelihood values (over a grid) from one data source.
}
\examples{
data("hmaLikelihoodList")
hmaLikEx <- hmaLikelihoodList[[1]]
\donttest{
plot(as.numeric(hmaLikEx[2,]) ~ as.numeric(names(hmaLikEx)))
}
}
\keyword{datasets}
