% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianMetaAnalysis.R
\name{loadCyclopsLibraryForJava}
\alias{loadCyclopsLibraryForJava}
\title{Load the Cyclops dynamic C++ library for use in Java}
\usage{
loadCyclopsLibraryForJava(
  file = system.file("libs", "Cyclops.so", package = "Cyclops")
)
}
\arguments{
\item{file}{The full system path to the \code{Cyclops.so} library}
}
\description{
Load the Cyclops dynamic C++ library for Markov chain Monte Carlo (MCMC) engine BEAST sampling.
}
