% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodApproximation.R
\name{prepareSccsIntervalData}
\alias{prepareSccsIntervalData}
\title{Prepare SCCS interval data for pooled analysis}
\usage{
prepareSccsIntervalData(sccsIntervalData, covariateId)
}
\arguments{
\item{sccsIntervalData}{An object of type \code{SccsIntervalData} as created using the \code{createSccsIntervalData}
function in the OHDSI \code{SelfControlledCaseSeries} package.}

\item{covariateId}{The ID of the covariate of interest, for which the estimate will be synthesized.
All other covariates will be considered nuisance variables.}
}
\value{
A tibble that can be used in the \code{computeBayesianMetaAnalysis()} function.
}
\description{
Prepare SCCS interval data for pooled analysis
}
