% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run simulation}
\usage{
runSimulation(settings)
}
\arguments{
\item{settings}{An object of type \code{simulationSettings}, created by the 
\code{\link{createSimulationSettings}} function.}
}
\value{
A object of class \code{simulation}..
}
\description{
Run simulation
}
\examples{
settings <- createSimulationSettings()
simulation <- runSimulation(settings)

}
