% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateMetaAnalysisWithNegativeControls}
\alias{simulateMetaAnalysisWithNegativeControls}
\title{Simulate survival data across a federated data network, with negative control outcomes as well.}
\usage{
simulateMetaAnalysisWithNegativeControls(
  meanExposureEffect = log(2),
  meanBias = 0.5,
  biasStd = 0.2,
  meanSiteEffect = 0,
  siteEffectStd = 0.1,
  mNegativeControls = 10,
  nSites = 10,
  sitePop = 2000,
  seed = 42,
  ...
)
}
\description{
Simulate survival data across a federated data network, with negative control outcomes as well.
}
