% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalMetaAnalysis.R
\name{summarizeChain}
\alias{summarizeChain}
\title{Utility function to summarize MCMC samples (posterior mean, median, HDI, std, etc.)}
\usage{
summarizeChain(chain, alpha = 0.05)
}
\arguments{
\item{chain}{A vector of posterior samples from MCMC.}

\item{alpha}{Alpha level for the credible interval.}
}
\description{
Utility function to summarize MCMC samples (posterior mean, median, HDI, std, etc.)
}
