% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianBiasCorrection.R
\name{biasCorrectionInference}
\alias{biasCorrectionInference}
\title{Bias Correction with Inference}
\usage{
biasCorrectionInference(
  likelihoodProfiles,
  biasDistribution = NULL,
  priorMean = 0,
  priorSd = 1,
  numsamps = 10000,
  thin = 10,
  doCorrection = TRUE,
  seed = 1,
  ...
)
}
\arguments{
\item{likelihoodProfiles}{A list of grid profile likelihoods regarding negative controls.}

\item{numsamps}{Total number of MCMC samples needed.}

\item{thin}{Thinning frequency: how many iterations before another sample is obtained?}

\item{seed}{Seed for the random number generator.}

\item{priorSds}{A two-dimensional vector with the standard deviation of the prior for
the average bias and the sd/scale parameter, respectively.}

\item{minNCs}{Minimum number of negative controls needed to fit a bias distribution;
default (also recommended): 5.}

\item{robust}{Whether or not to use a t-distribution model; default: FALSE.}
}
\value{
A dataframe with three columns and \code{numsamps} number of rows.
Column \code{mean} includes MCMC samples for the average bias,
\code{scale} for the sd/scale parameter,
and \code{bias} for predictive samples of the bias.
}
\description{
Perform Bayesian posterior inference regarding an outcome of interest with bias correction
using negative control analysis. There is an option to not perform bias correction so that
un-corrected results can be obtained.
}
\seealso{
\link{computeBayesianMetaAnalysis}
}
