% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectMetaAnalysis.R
\name{computeFixedEffectMetaAnalysis}
\alias{computeFixedEffectMetaAnalysis}
\title{Compute a fixed-effect meta-analysis}
\usage{
computeFixedEffectMetaAnalysis(data, alpha = 0.05)
}
\arguments{
\item{data}{A data frame containing either normal, skew-normal, custom parametric, or grid likelihood data. One row per database.}

\item{alpha}{The alpha (expected type I error) used for the confidence intervals.}
}
\description{
Compute a  fixed-effect meta-analysis.
}
\examples{
populations <- simulatePopulations()

fitModelInDatabase <- function(population) {
  cyclopsData <- Cyclops::createCyclopsData(Surv(time, y) ~ x + strata(stratumId), 
                                            data = population, modelType = "cox")
  cyclopsFit <- Cyclops::fitCyclopsModel(cyclopsData)
  approximation <-  approximateLikelihood(cyclopsFit, "x")
  return(approximation)
}
data <- lapply(populations, fitModelInDatabase)
data <- do.call("rbind", data)
computeFixedEffectMetaAnalysis(data)

}
