% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalMetaAnalysis.R
\name{computeHierarchicalMetaAnalysis}
\alias{computeHierarchicalMetaAnalysis}
\title{Compute a Bayesian random-effects hierarchical meta-analysis}
\usage{
computeHierarchicalMetaAnalysis(
  data,
  chainLength = 1100000,
  burnIn = 1e+05,
  subSampleFrequency = 100,
  alpha = 0.05,
  effectPriorStd = 1,
  globalExposureEffectPrior = c(0, 2),
  effectPrecisionPrior = c(1, 1),
  errorPrecisionPrior = c(1, 1),
  errorPrecisionStartValue = 1,
  includeSourceEffect = TRUE,
  includeExposureEffect = TRUE,
  seed = 1
)
}
\arguments{
\item{data}{A data frame containing either normal, skew-normal, custom parametric,
or grid likelihood data, with one row per database.}

\item{chainLength}{Number of MCMC iterations.}

\item{burnIn}{Number of MCMC iterations to consider as burn in.}

\item{subSampleFrequency}{Subsample frequency for the MCMC.}

\item{alpha}{The alpha (expected type I error) used for the credible intervals.}

\item{effectPriorStd}{Standard deviation for the average outcome / data-source effect.}

\item{globalExposureEffectPrior}{Prior mean and standard deviation for the global main effect.}

\item{effectPrecisionPrior}{Shape and scale for the gamma prior of the precision term in the
random effects model (normal) for individual outcome / data-source effects.}

\item{errorPrecisionPrior}{Shape and scale for the gamma prior of the precision term in the
normal model for random errors.}

\item{errorPrecisionStartValue}{Initial value for the error distribution's precision term.}

\item{includeSourceEffect}{Whether or not to consider the data-source-specific random effects. Default is TRUE.}

\item{includeExposureEffect}{Whether or not to estimate the main effect of interest. Default is TRUE.}

\item{seed}{Seed for the random number generator.}
}
\value{
A data frame with the point estimates, 95\% credible intervals and sample standard errors
for the (de-biased) global main effect, the average outcome effect, the average data source effect,
and precision of random errors.
Attributes of the data frame contain the MCMC trace and the detected approximation type.
}
\description{
Compute a Bayesian hierarchical meta-analysis (two-level model) to learn the global effect
with bias correction via negative control outcomes analysis.
Bayesian inference is performed using the Markov chain Monte Carlo (MCMC) engine BEAST.
Nnormal priors are used for the global effect, outcome-specific effects, and data-source-specific effects;
a half normal prior is used for the standard deviation; a gamma prior is used for the precision parameters.
}
\examples{
data("hmaLikelihoodList")
estimates = EvidenceSynthesis::computeHierarchicalMetaAnalysis(data = hmaLikelihoodList,
chainLength = 110000,
burnIn = 1e+04,
subSampleFrequency = 10,
seed = 666)

}
\seealso{
\link{approximateLikelihood}, \link{computeBayesianMetaAnalysis}
}
