% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidenceSynthesis.R
\docType{package}
\name{EvidenceSynthesis-package}
\alias{EvidenceSynthesis}
\alias{EvidenceSynthesis-package}
\title{EvidenceSynthesis: An R Package for Combining Evidence from Multiple Sources}
\description{
Routines for combining evidence and diagnostics across multiple 
   sources, such as multiple data sites in a distributed study. This includes functions
   for performing meta-analysis and forest plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/EvidenceSynthesis}
  \item Report bugs at \url{https://github.com/OHDSI/EvidenceSynthesis/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc A. Suchard \email{msuchard@ucla.edu}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
