% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidenceSynthesis.R
\docType{package}
\name{EvidenceSynthesis-package}
\alias{EvidenceSynthesis}
\alias{EvidenceSynthesis-package}
\title{EvidenceSynthesis: Synthesizing Causal Evidence in a Distributed Research Network}
\description{
Routines for combining causal effect estimates and study diagnostics across multiple data sites in a distributed study, without sharing patient-level data. Allows for normal and non-normal approximations of the data-site likelihood of the effect parameter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/EvidenceSynthesis/}
  \item \url{https://github.com/OHDSI/EvidenceSynthesis}
  \item Report bugs at \url{https://github.com/OHDSI/EvidenceSynthesis/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc A. Suchard \email{msuchard@ucla.edu}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
