% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotMcmcTrace}
\alias{plotMcmcTrace}
\title{Plot MCMC trace}
\usage{
plotMcmcTrace(
  estimate,
  showEstimate = TRUE,
  dataCutoff = 0.01,
  fileName = NULL
)
}
\arguments{
\item{estimate}{An object as generated using the \code{\link[=computeBayesianMetaAnalysis]{computeBayesianMetaAnalysis()}} function.}

\item{showEstimate}{Show the parameter estimates (mode) and 95 percent confidence intervals?}

\item{dataCutoff}{This fraction of the data at both tails will be removed.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \link[ggplot2:ggsave]{ggplot2::ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A Ggplot object. Use the \link[ggplot2:ggsave]{ggplot2::ggsave} function to save to file.
}
\description{
Plot MCMC trace
}
