% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{plotMetaAnalysisForest}
\alias{plotMetaAnalysisForest}
\title{Perform a meta-analysis and create a forest plot}
\usage{
plotMetaAnalysisForest(logRr, logLb95Ci, logUb95Ci, labels,
  xLabel = "Relative risk", limits = c(0.1, 10), fileName = NULL)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale.}

\item{logLb95Ci}{The lower bound of the 95 percent confidence interval on the log scale.}

\item{logUb95Ci}{The upper bound of the 95 percent confidence interval on the log scale.}

\item{labels}{A vector containing the labels for the various estimates.}

\item{xLabel}{The label on the x-axis: the name of the effect estimate.}

\item{limits}{The limits of the effect size axis.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A Ggplot object. Use the \code{ggsave} function to save to file.
}
\description{
\code{plotMetaAnalysisForest} performs a meta-analysis and creates a forest plot of effect size estimates.
}
\details{
Creates a forest plot of effect size estimates (ratios).
}
\examples{
plotMetaAnalysisForest(logRr = c(0, 0.2, -0.2, 0, 0.2, -0.2),
                       logLb95Ci = c(-0.2, -0.2, -0.6, -0.2, -0.2, -0.6),
                       logUb95Ci = c(0.2, 0.6, 0.2, 0.2, 0.6, 0.2),
                       labels = c("Site A", "Site B", "Site C", "Site D", "Site E", "Site F"))

}
