% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodApproximation.R
\name{skewNormal}
\alias{skewNormal}
\title{The skew normal function to approximate a log likelihood function}
\usage{
skewNormal(x, mu, sigma, alpha)
}
\arguments{
\item{x}{The log(hazard ratio) for which to approximate the log likelihood.}

\item{mu}{The position parameter.}

\item{sigma}{The scale parameter.}

\item{alpha}{The skew parameter.}
}
\value{
The approximate log likehood for the given x.
}
\description{
The skew normal function to approximate a log likelihood function
}
